requirements.txt
APScheduler==3.7.0
CurrencyConverter==0.16.5
nltk==3.5
pandas==1.0.5
PyDictionary==2.0.1
PyMySQL==1.0.2
python-amazon-sp-api==0.3.3
requests==2.25.1
spacy==2.3.5
stripe==2.56.0
loguru==0.5.3
fastapi==0.63.0
uvicorn==0.13.4
uvloop==0.15.2
APScheduler==3.7.0
